refactor nested objects

+++ app/src/main/java/com/leeloo/esist/db/dao/AttendanceDao.kt
@@ -13,6 +13,6 @@ interface AttendanceDao {
     fun insertAttendance(attendance: AttendanceEntity): Long
 
     @Delete
-    fun removeAttendance(attendance: AttendanceEntity): Long
+    fun removeAttendance(attendance: AttendanceEntity): Int
 
 }
